+++ packages/postcss-tape/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Tape
 
+### 1.2.0 (December 14, 2016)
+
+- Added: Expected `error` option
+- Added: `after` command to run option
+
 ### 1.1.0 (December 8, 2016)
 
 - Added: Support `plugins#process` with separate arguments